pyproject.toml
[tool.poetry]
name = "docker-swarm-ctl"
version = "0.1.0"
description = ""
authors = [" vilosource <vilosource@users.noreply.github.com>"]
readme = "README.md"
package-mode = false



[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.14"
uvicorn = "^0.35.0"
python-dotenv = "^1.1.1"
requests = "^2.32.4"
docker = "^7.1.0"
pyyaml = "^6.0.2"
pydantic-settings = "^2.10.1"


[tool.poetry.scripts]
dsctl = "dsctl_cli.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
